requirements.txt
asgiref==3.8.1
django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
drf-nested-routers==0.94.1
sqlparse==0.5.3
